+++ popup.js
@@ -2,7 +2,7 @@ chrome.windows.create({
   url: 'monitor.html',
   type: 'popup',
   width: 360,
-  height: 420,
+  height: 560,
   left: 100,
   top: 100
 }, (newWindow) => {